requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
Markdown==3.6
mysqlclient==2.2.4
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.26.3
